+++ apps/meteor/client/sidebar/header/actions/hooks/useAppsItems.tsx
@@ -3,7 +3,7 @@ import type { GenericMenuItemProps } from '@rocket.chat/ui-client';
 import { useTranslation, useRoute, usePermission } from '@rocket.chat/ui-contexts';
 import React from 'react';
 
-import { useUserDropdownAppsActionButtons } from '../../../../hooks/useAppActionButtons';
+import { useUserDropdownAppsActionButtons } from '../../../../hooks/useUserDropdownAppsActionButtons';
 import { useAppRequestStats } from '../../../../views/marketplace/hooks/useAppRequestStats';
 
 /**