+++ packages/freeswitch/src/connect.ts
@@ -6,7 +6,12 @@ import { logger } from './logger';
 
 const defaultPassword = 'ClueCon';
 
-export async function connect(options?: { host?: string; port?: number; password?: string }): Promise<FreeSwitchResponse> {
+export type EventNames = Parameters<FreeSwitchResponse['event_json']>;
+
+export async function connect(
+	options?: { host?: string; port?: number; password?: string },
+	customEventNames: EventNames = [],
+): Promise<FreeSwitchResponse> {
 	const host = options?.host ?? '127.0.0.1';
 	const port = options?.port ?? 8021;
 	const password = options?.password ?? defaultPassword;
@@ -26,7 +31,7 @@ export async function connect(options?: { host?: string; port?: number; password
 					await currentCall.onceAsync('freeswitch_auth_request', 20_000, 'FreeSwitchClient expected authentication request');
 					await currentCall.auth(password);
 					currentCall.auto_cleanup();
-					await currentCall.event_json('CHANNEL_EXECUTE_COMPLETE', 'BACKGROUND_JOB');
+					await currentCall.event_json('CHANNEL_EXECUTE_COMPLETE', 'BACKGROUND_JOB', ...customEventNames);
 				} catch (error) {
 					logger.error('FreeSwitchClient: connect error', error);
 					reject(error);