+++ packages/ui-video-conf/CHANGELOG.md
@@ -1,5 +1,35 @@
 # @rocket.chat/ui-video-conf
 
+## 13.0.0
+
+### Patch Changes
+
+- <details><summary>Updated dependencies []:</summary>
+
+  - @rocket.chat/ui-contexts@13.0.0
+  - @rocket.chat/ui-avatar@9.0.0
+  </details>
+
+## 13.0.0-rc.3
+
+### Patch Changes
+
+- <details><summary>Updated dependencies []:</summary>
+
+  - @rocket.chat/ui-contexts@13.0.0-rc.3
+  - @rocket.chat/ui-avatar@9.0.0-rc.3
+  </details>
+
+## 13.0.0-rc.2
+
+### Patch Changes
+
+- <details><summary>Updated dependencies []:</summary>
+
+  - @rocket.chat/ui-contexts@13.0.0-rc.2
+  - @rocket.chat/ui-avatar@9.0.0-rc.2
+  </details>
+
 ## 13.0.0-rc.1
 
 ### Patch Changes